heaters: sensors moved to own module
         Now thermistors are defined as a module with own section to
         make configuration simple. Sensor section is referenced
         in extruder or heater_bed section to get temperature read from
         correct sensor.
         This is also preparation for coming thermocouple/RTD support.
