+++ config/avrsim.cfg
@@ -36,6 +36,18 @@ position_min: 0.1
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog7
+min_temp: 0
+max_temp: 210
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog0
+min_temp: 0
+max_temp: 110
+
 [extruder]
 # Pins: PC3, PC2
 step_pin: ar19
@@ -44,24 +56,18 @@ enable_pin: ar25
 step_distance: .004242
 nozzle_diameter: 0.500
 filament_diameter: 3.500
+sensor: hotend
 heater_pin: ar4
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog7
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
 min_extrude_temp: 0
-max_temp: 210
 
 [heater_bed]
+sensor: bed
 heater_pin: ar3
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog0
 control: watermark
-min_temp: 0
-max_temp: 110
 
 [fan]
 pin: ar14
