+++ config/example-corexy.cfg
@@ -40,6 +40,18 @@ endstop_pin: ^ar18
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: analog13
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog14
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: ar26
 dir_pin: ar28
@@ -47,23 +59,17 @@ enable_pin: !ar24
 step_distance: .0022
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: ar10
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: analog13
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: ar8
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog14
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: ar9
