+++ config/example-multi-mcu.cfg
@@ -51,6 +51,18 @@ endstop_pin: ^zboard:ar18
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: auxboard:analog13
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: auxboard:analog14
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: auxboard:ar26
 dir_pin: auxboard:ar28
@@ -58,23 +70,17 @@ enable_pin: !auxboard:ar24
 step_distance: .002
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: auxboard:ar10
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: auxboard:analog13
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: auxboard:ar8
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: auxboard:analog14
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: auxboard:ar9
