+++ config/example.cfg
@@ -82,6 +82,39 @@ endstop_pin: ^ar18
 position_endstop: 0.5
 position_max: 200
 
+# The sensor section is used to describe sensor parameters for heaters
+# This section name is then referenced in extruders section.
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+#   Type of sensor - this may be "EPCOS 100K B57560G104F", "ATC
+#   Semitec 104GT-2", "NTC 100K beta 3950", or "AD595". This parameter
+#   must be provided.
+sensor_pin: analog13
+#   Analog input pin connected to the sensor. This parameter must be
+#   provided.
+#pullup_resistor: 4700
+#   The resistance (in ohms) of the pullup attached to the
+#   thermistor. This parameter is only valid when the sensor is a
+#   thermistor. The default is 4700 ohms.
+#adc_voltage: 5.0
+#   The ADC comparison voltage. This parameter is only valid when the
+#   sensor is an AD595. The default is 5 volts.
+min_temp: 0
+#   Minimum temperature in Celsius (mcu will shutdown if not
+#   met). This parameter must be provided.
+max_temp: 210
+#   Maximum temperature (mcu will shutdown if temperature is above
+#   this value). This parameter must be provided.
+
+# This section is similar to above but is used for bed heater
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog14
+min_temp: 0
+max_temp: 110
+
+
+
 # The extruder section is used to describe both the stepper
 # controlling the printer extruder and the heater parameters for the
 # nozzle. The stepper configuration has the same settings as the
@@ -134,6 +167,9 @@ filament_diameter: 3.500
 #   non-zero. The default is 0.010 (10 milliseconds).
 #
 # The remaining variables describe the extruder heater.
+sensor: hotend
+#   This parameter is reference to sensor section. Heater temperature
+#   is read by using referenced sensor.
 heater_pin: ar10
 #   PWM output pin controlling the heater. This parameter must be
 #   provided.
@@ -144,20 +180,6 @@ heater_pin: ar10
 #   allow the pin to be enabled for no more than half the time. This
 #   setting may be used to limit the total power output (over extended
 #   periods) to the heater. The default is 1.0.
-sensor_type: EPCOS 100K B57560G104F
-#   Type of sensor - this may be "EPCOS 100K B57560G104F", "ATC
-#   Semitec 104GT-2", "NTC 100K beta 3950", or "AD595". This parameter
-#   must be provided.
-sensor_pin: analog13
-#   Analog input pin connected to the sensor. This parameter must be
-#   provided.
-#pullup_resistor: 4700
-#   The resistance (in ohms) of the pullup attached to the
-#   thermistor. This parameter is only valid when the sensor is a
-#   thermistor. The default is 4700 ohms.
-#adc_voltage: 5.0
-#   The ADC comparison voltage. This parameter is only valid when the
-#   sensor is an AD595. The default is 5 volts.
 control: pid
 #   Control algorithm (either pid or watermark). This parameter must
 #   be provided.
@@ -185,27 +207,18 @@ pid_Kd: 114
 #min_extrude_temp: 170
 #   The minimum temperature (in Celsius) at which extruder move
 #   commands may be issued. The default is 170 Celsius.
-min_temp: 0
-#   Minimum temperature in Celsius (mcu will shutdown if not
-#   met). This parameter must be provided.
-max_temp: 210
-#   Maximum temperature (mcu will shutdown if temperature is above
-#   this value). This parameter must be provided.
 
 # The heater_bed section describes a heated bed (if present - omit
 # section if not present).
 [heater_bed]
+sensor: bed
 heater_pin: ar8
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog14
 control: watermark
 #max_delta: 2.0
 #   On 'watermark' controlled heaters this is the number of degrees in
 #   Celsius above the target temperature before disabling the heater
 #   as well as the number of degrees below the target before
 #   re-enabling the heater. The default is 2 degrees Celsius.
-min_temp: 0
-max_temp: 110
 
 # Print cooling fan (omit section if fan not present).
 [fan]
