+++ config/generic-cramps.cfg
@@ -39,6 +39,18 @@ endstop_pin: ^P9_13
 position_endstop: 0
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: P9_36
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: P9_33
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: P9_16
 dir_pin: P9_12
@@ -46,23 +58,17 @@ enable_pin: !P9_14
 step_distance: .002
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: P9_15
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: P9_36
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: P8_11
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: P9_33
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: P9_41
