+++ config/generic-mini-rambo.cfg
@@ -33,6 +33,19 @@ endstop_pin: ^PB4
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF0
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF2
+min_temp: 0
+max_temp: 130
+
+
 [extruder]
 step_pin: PC3
 dir_pin: PL6
@@ -40,23 +53,17 @@ enable_pin: !PA4
 step_distance: .002
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PE5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF0
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: PG5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF2
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: PH5
