+++ config/generic-printrboard.cfg
@@ -37,6 +37,18 @@ endstop_pin: ^PE4
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF1
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF0
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: PA6
 dir_pin: PA7
@@ -44,23 +56,17 @@ enable_pin: !PC3
 step_distance: .002
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PC5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF1
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: PC4
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF0
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: PC6
