+++ config/generic-replicape.cfg
@@ -99,6 +99,18 @@ max_accel: 3000
 max_z_velocity: 25
 max_z_accel: 30
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: host:analog4
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: host:analog6
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: P9_12
 dir_pin: P8_15
@@ -106,23 +118,17 @@ enable_pin: replicape:stepper_e_enable
 step_distance: .002
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: replicape:power_e
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: host:analog4
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: replicape:power_hotbed
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: host:analog6
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: replicape:power_fan0
