+++ config/printer-anet-a8-2017.cfg
@@ -40,6 +40,18 @@ position_endstop: 0.5
 position_max: 240
 homing_speed: 20
 
+[sensor hotend]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: PA7
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: PA6
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: PB1
 dir_pin: PB0
@@ -47,23 +59,17 @@ enable_pin: !PD6
 step_distance: .0105
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PD5
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: PA7
 control: pid
 pid_Kp: 2.151492
 pid_Ki: 0.633897
 pid_Kd: 230.042965
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: PD4
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: PA6
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: PB4
