+++ config/printer-anycubic-i3-mega-2017.cfg
@@ -44,6 +44,18 @@ enable_pin: !ar30
 step_distance: .0025
 endstop_pin: ^!ar43
 
+[sensor hotend]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: analog13
+min_temp: 0
+max_temp: 245
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog14
+min_temp: 0
+max_temp: 110
+
 [extruder]
 step_pin: ar26
 dir_pin: ar28
@@ -51,29 +63,23 @@ enable_pin: !ar24
 step_distance: .010799
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: ar10
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: analog13
 control: pid
 pid_Kp: 15.717
 pid_Ki: 0.569
 pid_Kd: 108.451
-min_temp: 0
-max_temp: 245
 
 [heater_fan extruder_fan]
 pin: ar44
 
 [heater_bed]
+sensor: bed
 heater_pin: ar8
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog14
 control: pid
 pid_Kp: 74.883
 pid_Ki: 1.809
 pid_Kd: 775.038
-min_temp: 0
-max_temp: 110
 
 [fan]
 pin: ar9
