+++ config/printer-creality-cr10-2017.cfg
@@ -40,6 +40,18 @@ endstop_pin: ^PC4
 position_endstop: 0.0
 position_max: 400
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PA7
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PA6
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: PB1
 dir_pin: !PB0
@@ -47,26 +59,20 @@ enable_pin: !PD6
 step_distance: 0.010526
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PD5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PA7
 control: pid
 pid_Kp: 22.57
 pid_Ki: 1.72
 pid_Kd: 73.96
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: PD4
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PA6
 control: pid
 pid_Kp: 426.68
 pid_Ki: 78.92
 pid_Kd: 576.71
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: PB4
