+++ config/printer-creality-cr10s-2017.cfg
@@ -32,6 +32,18 @@ endstop_pin: ^ar18
 position_endstop: 0.5
 position_max: 200
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog13
+min_temp: 0
+max_temp: 250
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: analog14
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: ar26
 dir_pin: ar28
@@ -39,26 +51,20 @@ enable_pin: !ar24
 step_distance: .010526
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: ar10
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog13
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 250
 
 [heater_bed]
+sensor: bed
 heater_pin: ar8
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: analog14
 control: pid
 pid_Kp: 690.34
 pid_Ki: 111.47
 pid_Kd: 1068.83
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: ar9
