+++ config/printer-lulzbot-taz6-2017.cfg
@@ -37,6 +37,24 @@ position_min: -1.5
 position_max: 270
 homing_speed: 1
 
+[sensor hotend1]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: PF0
+min_temp: 0
+max_temp: 300
+
+#[sensor hotend2]
+#sensor_type: ATC Semitec 104GT-2
+#sensor_pin: PF1
+#min_temp: 0
+#max_temp: 300
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF2
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: PC3
 dir_pin: !PL6
@@ -44,32 +62,26 @@ enable_pin: !PA4
 step_distance: 0.001182
 nozzle_diameter: 0.400
 filament_diameter: 2.920
+sensor: hotend1
 heater_pin: PH6
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: PF0
 control: pid
 pid_Kp: 28.79
 pid_Ki: 1.91
 pid_Kd: 108.51
-min_temp: 0
-max_temp: 300
 min_extrude_temp: 140
 
 #[extruder1]
 #step_pin: PC4
 #dir_pin: PL7
 #enable_pin: !PA3
+#sensor: hotend2
 #heater_pin: PH4
-#sensor_pin: PF1
 #...
 
 [heater_bed]
+sensor: bed
 heater_pin: PE5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF2
 control: watermark
-min_temp: 0
-max_temp: 130
 
 [fan]
 pin: PH5
