+++ config/printer-makergear-m2-2012.cfg
@@ -40,6 +40,18 @@ homing_retract_dist: 2.0
 homing_stepper_phases: 32
 homing_endstop_accuracy: .070
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF0
+min_temp: 0
+max_temp: 210
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF2
+min_temp: 0
+max_temp: 100
+
 [extruder]
 step_pin: PC3
 dir_pin: PL6
@@ -48,23 +60,17 @@ step_distance: .004242
 nozzle_diameter: 0.350
 filament_diameter: 1.750
 pressure_advance: 0.07
+sensor: hotend
 heater_pin: PH6
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF0
 control: pid
 pid_Kp: 7.0
 pid_Ki: 0.1
 pid_Kd: 12
-min_temp: 0
-max_temp: 210
 
 [heater_bed]
+sensor: bed
 heater_pin: PE5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF2
 control: watermark
-min_temp: 0
-max_temp: 100
 
 [fan]
 pin: PH5
