+++ config/printer-seemecnc-rostock-max-v2-2015.cfg
@@ -28,6 +28,18 @@ enable_pin: !PA5
 step_distance: .0125
 endstop_pin: ^PC7
 
+[sensor hotend]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: PF0
+min_temp: 0
+max_temp: 300
+
+[sensor bed]
+sensor_type: ATC Semitec 104GT-2
+sensor_pin: PF2
+min_temp: 0
+max_temp: 130
+
 [extruder]
 step_pin: PC3
 dir_pin: !PL6
@@ -35,26 +47,20 @@ enable_pin: !PA4
 step_distance: .010793
 nozzle_diameter: 0.500
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PH6
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: PF0
 control: pid
 pid_Kp: 20.9700
 pid_Ki: 1.3400
 pid_Kd: 80.5600
-min_temp: 0
-max_temp: 300
 
 [heater_bed]
+sensor: bed
 heater_pin: PE5
-sensor_type: ATC Semitec 104GT-2
-sensor_pin: PF2
 control: pid
 pid_Kp: 46.510
 pid_Ki: 1.040
 pid_Kd: 500.000
-min_temp: 0
-max_temp: 300
 
 [fan]
 pin: PH5
