+++ config/printer-tronxy-x5s-2017.cfg
@@ -40,6 +40,18 @@ endstop_pin: ^!PC4
 position_endstop: 0.5
 position_max: 400
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PA7
+min_temp: 0
+max_temp: 275
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PA6
+min_temp: 0
+max_temp: 150
+
 [extruder]
 step_pin: PB1
 dir_pin: PB0
@@ -47,23 +59,17 @@ enable_pin: !PD6
 step_distance: .0111
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PD5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PA7
 control: pid
 pid_Kp: 22.2
 pid_Ki: 1.08
 pid_Kd: 114
-min_temp: 0
-max_temp: 275
 
 [heater_bed]
+sensor: bed
 heater_pin: PD4
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PA6
 control: watermark
-min_temp: 0
-max_temp: 150
 
 [fan]
 pin: PB4
