+++ config/printer-wanhao-duplicator-i3-plus-2017.cfg
@@ -36,6 +36,18 @@ endstop_pin: ^!PA1
 position_endstop: 0.5
 position_max: 180
 
+[sensor hotend]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PF1
+min_temp: 0
+max_temp: 260
+
+[sensor bed]
+sensor_type: EPCOS 100K B57560G104F
+sensor_pin: PK6
+min_temp: 0
+max_temp: 110
+
 [extruder]
 step_pin: PF4
 dir_pin: PF5
@@ -43,26 +55,20 @@ enable_pin: !PF3
 step_distance: 0.010417
 nozzle_diameter: 0.400
 filament_diameter: 1.750
+sensor: hotend
 heater_pin: PG5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PF1
 control: pid
 pid_Kp: 30.850721
 pid_Ki: .208175
 pid_Kd: 192.298728
-min_temp: 0
-max_temp: 260
 
 [heater_bed]
+sensor: bed
 heater_pin: PE5
-sensor_type: EPCOS 100K B57560G104F
-sensor_pin: PK6
 control: pid
 pid_Kp: 64.095903
 pid_Ki: 1.649830
 pid_Kd: 622.531455
-min_temp: 0
-max_temp: 110
 
 [fan]
 pin: PE3
